Branch/feature cbh (#1)
* 1)Added reusable step in feature file for clicking  on link TV, Appliances, Electronics

2)Added reusable Wait method to avoid immediate termination of browser.

* 1)Added reusable step in feature file for clicking  on sub link Televisions

2)Added reusable step for Scrolls to Brand area.

* 1)Added reusable step in feature file for clicking Samsung Checkbox

2)Added reusable Sort by Price High to low step

* Added step to click on 2nd highest Price Item

* Added step For Switching the driver to new window.

* End to End Basic Flow is Done As per Assignment by CBH, Refinements and Enhancements are Pending.

* (Small changes as Samsung check box may not be available always in Brands area)

+++ src/main/java/com/selenium/cbh/Utilities.java
@@ -89,6 +89,8 @@ public class Utilities extends base{
         return isExists;
     }
 
+
+    //This one is for Android , Placeholder
     public static boolean isByExists(By by, AndroidDriver<AndroidElement> AD) {
         boolean isExists = true;
         try {
@@ -110,8 +112,8 @@ public class Utilities extends base{
     }
 
 
-    public static int minutesToMilliSecond(int min) {
-        int sec;
+    public static long minutesToMilliSecond(long min) {
+        long sec;
 
         sec = min * 60 * 1000;
